update - double to int for montant

+++ virement.h
@@ -5,7 +5,7 @@ typedef struct
 {
   int num_destinataire;
   int num_expediteur;
-  double montant;
+  int montant;
 } Virement;
 
 #endif
